+++ project/automaton_transition.py
